+++ Assets/Proto/Scripts/Shooting/ShootingTarget/StaticShootingTarget/StaticShootingTarget.cs
@@ -62,7 +62,7 @@ namespace AppleShooterProto{
 			public new class ConstArg: AbsShootingTarget.ConstArg, IConstArg{
 				public ConstArg(
 					int index,
-					float health,
+					int health,
 					Color defaultColor,
 					UnityBase.IBellCurve healthBellCurve,
 					float heatBonus,
